+++ src/chocolatey/infrastructure.app/ApplicationParameters.cs
@@ -53,7 +53,7 @@ namespace chocolatey.infrastructure.app
         public static readonly string ChocolateyCommunityFeedPushSource = "https://chocolatey.org/";
         public static readonly string UserAgent = "Chocolatey Command Line";
         public static readonly string RegistryValueInstallLocation = "InstallLocation";
-        public static readonly string RollbackPackageSuffix = "._.previous";
+        public static readonly string RollbackPackageSuffix = "._.old";
 
         /// <summary>
         ///   Default is 45 minutes