+++ src/chocolatey/infrastructure.app/services/PowershellService.cs
@@ -76,11 +76,15 @@ namespace chocolatey.infrastructure.app.services
             return run_action(configuration, packageResult, CommandNameType.uninstall);
         }
 
-        public string wrap_command_with_module(string command)
+        public string wrap_script_with_module(string script)
         {
             var installerModules = _fileSystem.get_files(ApplicationParameters.InstallLocation, "chocolateyInstaller.psm1", SearchOption.AllDirectories);
             var installerModule = installerModules.FirstOrDefault();
-            return "[System.Threading.Thread]::CurrentThread.CurrentCulture = '';[System.Threading.Thread]::CurrentThread.CurrentUICulture = ''; & import-module -name '{0}';$ErrorActionPreference = 'Stop'; {1}".format_with(installerModule, command);
+            // removed setting all errors to terminating. Will cause too
+            // many issues in existing packages, including upgrading
+            // Chocolatey from older POSH client due to log errors
+            //$ErrorActionPreference = 'Stop';
+            return "[System.Threading.Thread]::CurrentThread.CurrentCulture = '';[System.Threading.Thread]::CurrentThread.CurrentUICulture = ''; & import-module -name '{0}'; & '{1}'".format_with(installerModule, script);
         }
 
         public bool run_action(ChocolateyConfiguration configuration, PackageResult packageResult, CommandNameType command)
@@ -178,7 +182,7 @@ namespace chocolatey.infrastructure.app.services
                 {
                     installerRun = true;
                     var exitCode = PowershellExecutor.execute(
-                        wrap_command_with_module(chocoPowerShellScript),
+                        wrap_script_with_module(chocoPowerShellScript),
                         _fileSystem,
                         configuration.CommandExecutionTimeoutSeconds,
                         (s, e) =>