+++ CHANGELOG.md
@@ -16,8 +16,8 @@ We'll try to capture everything here that you should know about. Please call `ch
 
 ### BREAKING CHANGES
 
- * [Security] Perhaps the most significant breaking change is that security settings have pushed us to stop for confirmation before changing the state of the system on most commands. You can override with `-y`.
- * [Security] Moved default installation out of C:\Chocolatey, update environment settings  - see [#7](https://github.com/chocolatey/choco/issues/7)
+ * [Security] **Prompt for confirmation**: For security reasons, we now stop for confirmation before changing the state of the system on most commands. You can pass `-y` to confirm any prompts.
+ * [Security] If your default installation is still at `C:\Chocolatey`, this version will force a move to ProgramData and update the environment settings - see [#7](https://github.com/chocolatey/choco/issues/7)
  * **Configuration Breaking Changes:**
    1. You now have one config file to interact with in %ChocolateyInstall%\config - your user config is no longer valid and can be removed once you migrate settings to the config.
    2. The config will not be overwritten on upgrade.